Update HdfsFile.md
In the BaseSinkConfig.java code, there is a configuration key called “remote_user”, but it is not shown in the front-end documentation. This configuration is useful when you need to control read and write (access rights) to HDFS by specific users.

I have encountered the following errors in actual use: Caused by: `org.apache.hadoop.ipc.RemoteException: Permission denied:  user=root, access=WRITE, inode="/tmp/seatunnel/seatunnel"`. By setting remote_user="xxx" as the specified user in the HdfsFile of the Sink section, the above problems can be solved. 

Therefore, the configuration key should be displayed in the front-end documentation.